+++ src/main/java/insane96mcp/mobspropertiesrandomness/json/MPRMob.java
@@ -6,6 +6,7 @@ import insane96mcp.mobspropertiesrandomness.exception.InvalidJsonException;
 import insane96mcp.mobspropertiesrandomness.json.mobs.MPRCreeper;
 import insane96mcp.mobspropertiesrandomness.json.mobs.MPRGhast;
 import insane96mcp.mobspropertiesrandomness.json.mobs.MPRPhantom;
+import insane96mcp.mobspropertiesrandomness.json.utils.attribute.MPRMobAttribute;
 import insane96mcp.mobspropertiesrandomness.setup.Strings;
 import insane96mcp.mobspropertiesrandomness.utils.Logger;
 import insane96mcp.mobspropertiesrandomness.utils.MPRUtils;